--- a/.changeset/khaki-otters-smile.md
+++ b/.changeset/khaki-otters-smile.md
@@ -1,0 +1,5 @@
+---
+"@callstack/repack": patch
+---
+
+Improve performance of HMR