### Summary

This PR introduces a performance improvement for HMR in the `@callstack/repack` package.

### Changes:

- Modified `LoggerPlugin.ts` to improve performance. Specifically, an update has been made to the `stats.toJson` method. By adding `all: false` we disable retrieving other stats that were being needlessly retrieved by default.

### Performance Metrics:

#### HMR time:

| Metric | Before Change (ms) | After Change (ms) |
|---|---|---|
| Min | 482 | 433 |
| Max | 616 | 549 |
| Mean | 528.62 | 476.60 |
| Standard Deviation | 60.29 | 29.16 |

#### LoggerPlugin `done` hook time:

| Metric | Before Change (ms) | After Change (ms) |
|---|---|---|
| Min | 40 | 0.1 |
| Max | 146 | 1 |

---